Add reason property to GetToken events. Refactor code to reduce number of exported symbols
