+++ api-report/odsp-driver-definitions.api.md
@@ -58,9 +58,6 @@ export interface IFileEntry {
     resolvedUrl: IFluidResolvedUrl;
 }
 
-// @public (undocumented)
-export type InstrumentedStorageTokenFetcher = (options: TokenFetchOptions, name: string, alwaysRecordTokenFetchTelemetry?: boolean) => Promise<string | null>;
-
 // @public
 export interface IOdspError extends Omit<IDriverErrorBase, "errorType"> {
     // (undocumented)
