+++ packages/drivers/odsp-driver-definitions/src/tokenFetch.ts
@@ -88,8 +88,3 @@ export const isTokenFromCache = (tokenResponse: string | TokenResponse | null): 
  *  Enterprise represents user authenticated with M365 tenant account
  */
 export type IdentityType = "Consumer" | "Enterprise";
-
-export type InstrumentedStorageTokenFetcher = (
-    options: TokenFetchOptions,
-    name: string,
-    alwaysRecordTokenFetchTelemetry?: boolean) => Promise<string | null>;
