+++ packages/drivers/odsp-driver/src/index.ts
@@ -31,7 +31,3 @@ export * from "./odspDriverUrlResolver";
 
 // It's used by URL resolve code, but also has some public functions
 export * from "./odspFluidFileLink";
-
-// Wire format related
-export * from "./compactSnapshotParser";
-export * from "./ReadBufferUtils";
