+++ packages/drivers/odsp-driver/src/localOdspDriver/localOdspDocumentStorageManager.ts
@@ -9,7 +9,7 @@ import { ISummaryContext } from "@fluidframework/driver-definitions";
 import { UsageError } from "@fluidframework/driver-utils";
 import * as api from "@fluidframework/protocol-definitions";
 import { OdspDocumentStorageServiceBase } from "../odspDocumentStorageServiceBase";
-import { ISnapshotContents } from "../odspPublicUtils";
+import { ISnapshotContents } from "../contractsInternal";
 import { IOdspSnapshot } from "../contracts";
 import { convertOdspSnapshotToSnapshotTreeAndBlobs } from "../odspSnapshotParser";
 import { parseCompactSnapshotResponse } from "../compactSnapshotParser";
