+++ packages/drivers/odsp-driver/src/odspSummaryUploadManager.ts
@@ -8,7 +8,6 @@ import { assert, Uint8ArrayToString, unreachableCase } from "@fluidframework/com
 import { ISummaryContext } from "@fluidframework/driver-definitions";
 import { getGitType } from "@fluidframework/protocol-base";
 import * as api from "@fluidframework/protocol-definitions";
-import { InstrumentedStorageTokenFetcher } from "@fluidframework/odsp-driver-definitions";
 import { loggerToMonitoringContext, MonitoringContext, PerformanceEvent } from "@fluidframework/telemetry-utils";
 import {
     IOdspSummaryPayload,
@@ -21,6 +20,7 @@ import {
 import { EpochTracker } from "./epochTracker";
 import { getUrlAndHeadersWithAuth } from "./getUrlAndHeadersWithAuth";
 import { getWithRetryForTokenRefresh } from "./odspUtils";
+import { InstrumentedStorageTokenFetcher } from "./contractsInternal";
 
 /* eslint-disable max-len */
